move flappy jellyfish behavior into its module
